requirements.txt
streamlit==1.38.0
pdfplumber==0.11.4
python-dotenv==1.0.1
google-generativeai==0.7.2
pandas